De-duplicate the PVC based vms
Gather all PVC based VMs under vm-pvc directory. The vm resources are
located in the base kustomization directory, and an overlay per cluster:
k8s-regional, odr-regional, and odr-metro. The overlay overrides the
storage class and the network interface to match ODF and Ramen testing
environments.

When using the odr-regional and odr-metro versions in OpenShift console,
you need to specify the path like this:

    vm-pvc/odr-regional

Example layout:

    $ tree vm-pvc/
    vm-pvc/
    ├── base
    │   ├── kustomization.yaml
    │   ├── pvc.yaml
    │   ├── source.yaml
    │   ├── test_rsa.pub
    │   └── vm.yaml
    ├── k8s-regional
    │   └── kustomization.yaml
    ├── odr-metro
    │   └── kustomization.yaml
    └── odr-regional
        └── kustomization.yaml